chore: Rename soilproj app to project_management
