+++ docker-compose.dev.yml
@@ -10,11 +10,6 @@ services:
     depends_on:
       db:
         condition: service_healthy
-
-    depends_on:
-      db:
-        condition: service_healthy
-
   db:
     extends:
       file: docker-compose.base.yml
