+++ terraso_backend/config/settings.py
@@ -62,7 +62,7 @@ INSTALLED_APPS = [
     "apps.shared_data",
     "apps.story_map",
     "apps.notifications",
-    "apps.soilproj",
+    "apps.project_management",
 ]
 
 if DEBUG:
